pyproject.toml
[project]
name = "transcript-summarizer"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "David Armendariz", email = "david.adrian.a.p@gmail.com"},
]
dependencies = ["gradio>=5.10.0", "notebook>=7.3.2", "openai>=1.59.4", "pypdf>=5.1.0", "python-dotenv>=1.0.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
